refactored go tests
