+++ terratest/src/test/Gopkg.toml
@@ -1,7 +1,54 @@
+
+# Gopkg.toml example
+#
+# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
+# for detailed Gopkg.toml documentation.
+#
+# required = ["github.com/user/thing/cmd/thing"]
+# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
+#
+# [[constraint]]
+#   name = "github.com/user/project"
+#   version = "1.0.0"
+#
+# [[constraint]]
+#   name = "github.com/user/project2"
+#   branch = "dev"
+#   source = "github.com/myfork/project2"
+#
+# [[override]]
+#  name = "github.com/x/y"
+#  version = "2.4.0"
+
+
+[[constraint]]
+  name = "github.com/aws/aws-sdk-go"
+  version = "1.32.3"
+
+[[constraint]]
+  branch = "master"
+  name = "github.com/chromedp/cdproto"
+
+[[constraint]]
+  name = "github.com/chromedp/chromedp"
+  version = "0.5.3"
+
 [[constraint]]
   name = "github.com/gruntwork-io/terratest"
-  version = "v0.18.2"
-  
-[[override]]
-  name = "github.com/russross/blackfriday"
-  version = "1.5.2"
+  version = "0.28.5"
+
+[[constraint]]
+  name = "gopkg.in/yaml.v2"
+  version = "2.3.0"
+
+[[constraint]]
+  name = "k8s.io/api"
+  version = "0.18.3"
+
+[[constraint]]
+  name = "k8s.io/apimachinery"
+  version = "0.17.6"
+
+[[constraint]]
+  name = "k8s.io/client-go"
+  version = "0.18.3"
