+++ terratest/src/test/addons/DeploymentMetadata.go
@@ -1,4 +1,4 @@
-package pega
+package addons
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
