+++ terratest/src/test/addons/alb-ingress_test.go
@@ -3,19 +3,18 @@ package addons
 import (
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/apps/v1"
-	"test/common"
 	"testing"
 )
 
 func TestShouldNotContainAlbIngressIfDisabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled": "false",
 		}),
 	)
 
 	for _, i := range albIngressResources {
-		require.False(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.False(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -23,14 +22,14 @@ func TestShouldNotContainAlbIngressIfDisabled(t *testing.T) {
 }
 
 func TestAlbIngressShouldContainAllResources(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled": "true",
 		}),
 	)
 
 	for _, i := range albIngressResources {
-		require.True(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.True(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -38,8 +37,8 @@ func TestAlbIngressShouldContainAllResources(t *testing.T) {
 }
 
 func Test_checkSetAwsRegion(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":               "true",
 			"aws-alb-ingress-controller.autoDiscoverAwsRegion": "false",
 			"aws-alb-ingress-controller.awsRegion":             "YOUR_EKS_CLUSTER_REGION",
@@ -47,8 +46,8 @@ func Test_checkSetAwsRegion(t *testing.T) {
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -56,8 +55,8 @@ func Test_checkSetAwsRegion(t *testing.T) {
 }
 
 func Test_checkAutoDiscoverAwsRegion(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":               "true",
 			"aws-alb-ingress-controller.autoDiscoverAwsRegion": "true",
 			"aws-alb-ingress-controller.awsRegion":             "YOUR_EKS_CLUSTER_REGION",
@@ -65,8 +64,8 @@ func Test_checkAutoDiscoverAwsRegion(t *testing.T) {
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -74,8 +73,8 @@ func Test_checkAutoDiscoverAwsRegion(t *testing.T) {
 }
 
 func Test_checkSetAwsVpcID(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":              "true",
 			"aws-alb-ingress-controller.autoDiscoverAwsVpcID": "false",
 			"aws-alb-ingress-controller.awsVpcID":             "YOUR_EKS_CLUSTER_VPC_ID",
@@ -83,8 +82,8 @@ func Test_checkSetAwsVpcID(t *testing.T) {
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -92,8 +91,8 @@ func Test_checkSetAwsVpcID(t *testing.T) {
 }
 
 func Test_checkAutoDiscoverAwsVpcID(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":              "true",
 			"aws-alb-ingress-controller.autoDiscoverAwsVpcID": "true",
 			"aws-alb-ingress-controller.awsVpcID":             "YOUR_EKS_CLUSTER_VPC_ID",
@@ -101,8 +100,8 @@ func Test_checkAutoDiscoverAwsVpcID(t *testing.T) {
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -110,16 +109,16 @@ func Test_checkAutoDiscoverAwsVpcID(t *testing.T) {
 }
 
 func Test_checkSetClusterName(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":     "true",
 			"aws-alb-ingress-controller.clusterName": "YOUR_EKS_CLUSTER_NAME",
 		}),
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -127,8 +126,8 @@ func Test_checkSetClusterName(t *testing.T) {
 }
 
 func Test_checkSetAccessKey(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"aws-alb-ingress-controller.enabled":                        "true",
 			"aws-alb-ingress-controller.extraEnv.AWS_ACCESS_KEY_ID":     "YOUR_AWS_ACCESS_KEY_ID",
 			"aws-alb-ingress-controller.extraEnv.AWS_SECRET_ACCESS_KEY": "YOUR_AWS_SECRET_ACCESS_KEY",
@@ -136,8 +135,8 @@ func Test_checkSetAccessKey(t *testing.T) {
 	)
 
 	var deployment *v1.Deployment
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-aws-alb-ingress-controller",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -147,21 +146,21 @@ func Test_checkSetAccessKey(t *testing.T) {
 	require.Equal(t, "YOUR_AWS_SECRET_ACCESS_KEY", deployment.Spec.Template.Spec.Containers[0].Env[1].Value)
 }
 
-var albIngressResources = []common.SearchResourceOption{
+var albIngressResources = []SearchResourceOption{
 	{
-		Name: "release-name-aws-alb-ingress-controller",
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "ServiceAccount",
 	},
 	{
-		Name: "release-name-aws-alb-ingress-controller",
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "ClusterRole",
 	},
 	{
-		Name: "release-name-aws-alb-ingress-controller",
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "ClusterRoleBinding",
 	},
 	{
-		Name: "release-name-aws-alb-ingress-controller",
+		Name: "pega-aws-alb-ingress-controller",
 		Kind: "Deployment",
 	},
 }
