+++ terratest/src/test/addons/azure-ingress_test.go
@@ -4,19 +4,18 @@ import (
 	b64 "encoding/base64"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
-	"test/common"
 	"testing"
 )
 
 func TestShouldNotContainAzureIngressIfDisabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"ingress-azure.enabled": "false",
 		}),
 	)
 
 	for _, i := range azureIngressResources {
-		require.False(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.False(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -24,14 +23,14 @@ func TestShouldNotContainAzureIngressIfDisabled(t *testing.T) {
 }
 
 func TestAzureIngressShouldContainAllResources(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"ingress-azure.enabled": "true",
 		}),
 	)
 
 	for _, i := range azureIngressResources {
-		require.True(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.True(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -39,8 +38,8 @@ func TestAzureIngressShouldContainAllResources(t *testing.T) {
 }
 
 func TestSetValuesForAppGW(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"ingress-azure.enabled":              "true",
 			"ingress-azure.appgw.subscriptionId": "<YOUR.SUBSCRIPTION_ID>",
 			"ingress-azure.appgw.resourceGroup":  "<RESOURCE_GROUP_NAME>",
@@ -49,8 +48,8 @@ func TestSetValuesForAppGW(t *testing.T) {
 	)
 
 	var configMap *v1.ConfigMap
-	helmChart.Find(common.SearchResourceOption{
-		Name: "release-name-cm-ingress-azure",
+	helmChart.Find(SearchResourceOption{
+		Name: "pega-cm-ingress-azure",
 		Kind: "ConfigMap",
 	}, &configMap)
 
@@ -61,8 +60,8 @@ func TestSetValuesForAppGW(t *testing.T) {
 }
 
 func TestSetValuesForArmAuth(t *testing.T) {
-	helmChart := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChart := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"ingress-azure.enabled":            "true",
 			"ingress-azure.armAuth.type":       "servicePrincipal",
 			"ingress-azure.armAuth.secretJSON": b64.StdEncoding.EncodeToString([]byte("<SECRET_JSON_CREATED_USING_ABOVE_COMMAND>")),
@@ -70,7 +69,7 @@ func TestSetValuesForArmAuth(t *testing.T) {
 	)
 
 	var secret *v1.Secret
-	helmChart.Find(common.SearchResourceOption{
+	helmChart.Find(SearchResourceOption{
 		Name: "networking-appgw-k8s-azure-service-principal",
 		Kind: "Secret",
 	}, &secret)
@@ -78,29 +77,29 @@ func TestSetValuesForArmAuth(t *testing.T) {
 	require.Equal(t, "<SECRET_JSON_CREATED_USING_ABOVE_COMMAND>", string(secret.Data["armAuth.json"]))
 }
 
-var azureIngressResources = []common.SearchResourceOption{
+var azureIngressResources = []SearchResourceOption{
 	{
 		Name: "networking-appgw-k8s-azure-service-principal",
 		Kind: "Secret",
 	},
 	{
-		Name: "release-name-cm-ingress-azure",
+		Name: "pega-cm-ingress-azure",
 		Kind: "ConfigMap",
 	},
 	{
-		Name: "release-name-sa-ingress-azure",
+		Name: "pega-sa-ingress-azure",
 		Kind: "ServiceAccount",
 	},
 	{
-		Name: "release-name-ingress-azure",
+		Name: "pega-ingress-azure",
 		Kind: "ClusterRole",
 	},
 	{
-		Name: "release-name-ingress-azure",
+		Name: "pega-ingress-azure",
 		Kind: "ClusterRoleBinding",
 	},
 	{
-		Name: "release-name-ingress-azure",
+		Name: "pega-ingress-azure",
 		Kind: "Deployment",
 	},
 }
