+++ terratest/src/test/addons/efk_test.go
@@ -5,13 +5,12 @@ import (
 	"k8s.io/api/apps/v1beta2"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
-	"test/common"
 	"testing"
 )
 
 func TestShouldNotContainDeploy_EFKIfDisabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"elasticsearch.enabled":         "false",
 			"kibana.enabled":                "false",
 			"fluentd-elasticsearch.enabled": "false",
@@ -19,15 +18,15 @@ func TestShouldNotContainDeploy_EFKIfDisabled(t *testing.T) {
 	)
 
 	for _, i := range deployEfkResources {
-		require.False(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.False(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
 	}
 }
 func TestShouldDeploy_EFKContainAllResourcesIfEnabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"elasticsearch.enabled":         "true",
 			"kibana.enabled":                "true",
 			"fluentd-elasticsearch.enabled": "true",
@@ -35,7 +34,7 @@ func TestShouldDeploy_EFKContainAllResourcesIfEnabled(t *testing.T) {
 	)
 
 	for _, i := range deployEfkResources {
-		require.True(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.True(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -43,44 +42,44 @@ func TestShouldDeploy_EFKContainAllResourcesIfEnabled(t *testing.T) {
 }
 
 func Test_shouldBeFullnameOverrideForElasticsearch(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"elasticsearch.enabled": "true",
 		}),
 	)
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
 		Name: "elastic-search",
 		Kind: "ConfigMap",
 	}))
 }
 
 func Test_shouldBeElasticSearchUrlForKibana(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"kibana.enabled": "true",
 			"kibana.files.kibana.yml.elasticsearch.url": "http://elastic-search-client:9200",
 		}),
 	)
 
 	var configmap *v1.ConfigMap
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-kibana",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-kibana",
 		Kind: "ConfigMap",
 	}, &configmap)
 
 	require.Contains(t, configmap.Data["kibana.yml"], "http://elastic-search-client:9200")
 }
 func Test_shouldBeServiceExternalPortForKibana(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"kibana.enabled":              "true",
 			"kibana.service.externalPort": "80",
 		}),
 	)
 
 	var service *v1.Service
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-kibana",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-kibana",
 		Kind: "Service",
 	}, &service)
 
@@ -88,36 +87,36 @@ func Test_shouldBeServiceExternalPortForKibana(t *testing.T) {
 }
 
 func Test_shouldBeIngressEnabledForKibana(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"kibana.enabled":         "true",
 			"kibana.ingress.enabled": "true",
 		}),
 	)
 
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-kibana",
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-kibana",
 		Kind: "Ingress",
 	}))
 }
 
 func Test_shouldBeIngressDisabledForKibana(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"kibana.enabled":         "true",
 			"kibana.ingress.enabled": "false",
 		}),
 	)
 
-	require.False(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-kibana",
+	require.False(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-kibana",
 		Kind: "Ingress",
 	}))
 }
 
 func Test_shouldBeHostForIngressKibana(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"kibana.enabled":         "true",
 			"kibana.ingress.enabled": "true",
 			"kibana.ingress.hosts":   "{YOUR_WEB.KIBANA.EXAMPLE.COM}",
@@ -125,8 +124,8 @@ func Test_shouldBeHostForIngressKibana(t *testing.T) {
 	)
 
 	var ingress *v1beta1.Ingress
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-kibana",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-kibana",
 		Kind: "Ingress",
 	}, &ingress)
 
@@ -134,16 +133,16 @@ func Test_shouldBeHostForIngressKibana(t *testing.T) {
 
 }
 func Test_shouldBeHostForElasticsearch(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"fluentd-elasticsearch.enabled": "true",
 			"elasticsearch.host":            "elastic-search-client",
 		}),
 	)
 
 	var daemon *v1beta2.DaemonSet
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-fluentd-elasticsearch",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "DaemonSet",
 	}, &daemon)
 
@@ -151,16 +150,16 @@ func Test_shouldBeHostForElasticsearch(t *testing.T) {
 }
 
 func Test_shouldBeBufferChunkLimitForElasticsearch(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"fluentd-elasticsearch.enabled":    "true",
 			"elasticsearch.buffer_chunk_limit": "250M",
 		}),
 	)
 
 	var daemon *v1beta2.DaemonSet
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-fluentd-elasticsearch",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "DaemonSet",
 	}, &daemon)
 
@@ -168,33 +167,33 @@ func Test_shouldBeBufferChunkLimitForElasticsearch(t *testing.T) {
 }
 
 func Test_shouldBeBufferQueueLimitForElasticsearch(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"fluentd-elasticsearch.enabled":    "true",
 			"elasticsearch.buffer_queue_limit": "30",
 		}),
 	)
 
 	var daemon *v1beta2.DaemonSet
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-fluentd-elasticsearch",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "DaemonSet",
 	}, &daemon)
 
 	require.Equal(t, "30", daemon.Spec.Template.Spec.Containers[0].Env[5].Value)
 }
 
-var deployEfkResources = []common.SearchResourceOption{
+var deployEfkResources = []SearchResourceOption{
 	{
-		Name: "release-name-kibana",
+		Name: "pega-kibana",
 		Kind: "ConfigMap",
 	},
 	{
-		Name: "release-name-kibana",
+		Name: "pega-kibana",
 		Kind: "Service",
 	},
 	{
-		Name: "release-name-kibana",
+		Name: "pega-kibana",
 		Kind: "Deployment",
 	},
 	{
@@ -234,23 +233,23 @@ var deployEfkResources = []common.SearchResourceOption{
 		Kind: "Deployment",
 	},
 	{
-		Name: "release-name-fluentd-elasticsearch",
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "ConfigMap",
 	},
 	{
-		Name: "release-name-fluentd-elasticsearch",
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "ServiceAccount",
 	},
 	{
-		Name: "release-name-fluentd-elasticsearch",
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "ClusterRole",
 	},
 	{
-		Name: "release-name-fluentd-elasticsearch",
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "ClusterRoleBinding",
 	},
 	{
-		Name: "release-name-fluentd-elasticsearch",
+		Name: "pega-fluentd-elasticsearch",
 		Kind: "DaemonSet",
 	},
 }
