+++ terratest/src/test/addons/traefik_test.go
@@ -4,19 +4,18 @@ import (
 	"github.com/stretchr/testify/require"
 	v12 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
-	"test/common"
 	"testing"
 )
 
 func Test_shouldNotContainTraefikResourcesWhenDisabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled": "false",
 		}),
 	)
 
 	for _, i := range traefikResources {
-		require.False(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.False(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -24,14 +23,14 @@ func Test_shouldNotContainTraefikResourcesWhenDisabled(t *testing.T) {
 }
 
 func Test_shouldContainTraefikResourcesWhenEnabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled": "true",
 		}),
 	)
 
 	for _, i := range traefikResources {
-		require.True(t, helmChartParser.Contains(common.SearchResourceOption{
+		require.True(t, helmChartParser.Contains(SearchResourceOption{
 			Name: i.Name,
 			Kind: i.Kind,
 		}))
@@ -39,16 +38,16 @@ func Test_shouldContainTraefikResourcesWhenEnabled(t *testing.T) {
 }
 
 func Test_shouldBeAbleToSetUpServiceTypeAsLoadBalancer(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":     "true",
 			"traefik.serviceType": "LoadBalancer",
 		}),
 	)
 
 	var service *v1.Service
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Service",
 	}, &service)
 
@@ -57,16 +56,16 @@ func Test_shouldBeAbleToSetUpServiceTypeAsLoadBalancer(t *testing.T) {
 }
 
 func Test_shouldBeAbleToSetUpServiceTypeAsNodePort(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":     "true",
 			"traefik.serviceType": "NodePort",
 		}),
 	)
 
 	var service *v1.Service
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Service",
 	}, &service)
 
@@ -77,83 +76,83 @@ func Test_shouldBeAbleToSetUpServiceTypeAsNodePort(t *testing.T) {
 }
 
 func Test_hasRoleWhenRbacEnabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":      "true",
 			"traefik.rbac.enabled": "true",
 		}),
 	)
 
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "ClusterRole",
 	}))
 
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "ServiceAccount",
 	}))
 
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "ClusterRoleBinding",
 	}))
 }
 
 func Test_noRoleWhenRbacDisabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":      "true",
 			"traefik.rbac.enabled": "false",
 		}),
 	)
 
-	require.False(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	require.False(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "ClusterRole",
 	}))
 }
 
 func Test_hasSecretWhenSSLEnabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":     "true",
 			"traefik.ssl.enabled": "true",
 		}),
 	)
 
 	var deployment v12.Deployment
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	}, &deployment)
 
-	require.True(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik-default-cert",
+	require.True(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik-default-cert",
 		Kind: "Secret",
 	}))
 
 	require.Equal(t, "ssl", deployment.Spec.Template.Spec.Volumes[1].Name)
-	require.Equal(t, "release-name-traefik-default-cert", deployment.Spec.Template.Spec.Volumes[1].Secret.SecretName)
+	require.Equal(t, "pega-traefik-default-cert", deployment.Spec.Template.Spec.Volumes[1].Secret.SecretName)
 }
 
 func Test_hasNoSecretWhenSSLEnabled(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":     "true",
 			"traefik.ssl.enabled": "false",
 		}),
 	)
 
-	require.False(t, helmChartParser.Contains(common.SearchResourceOption{
-		Name: "release-name-traefik-default-cert",
+	require.False(t, helmChartParser.Contains(SearchResourceOption{
+		Name: "pega-traefik-default-cert",
 		Kind: "Secret",
 	}))
 }
 
 func Test_checkResourceRequests(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":                   "true",
 			"traefik.resources.requests.cpu":    "300m",
 			"traefik.resources.requests.memory": "300Mi",
@@ -161,8 +160,8 @@ func Test_checkResourceRequests(t *testing.T) {
 	)
 
 	var deployment v12.Deployment
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -171,8 +170,8 @@ func Test_checkResourceRequests(t *testing.T) {
 }
 
 func Test_checkResourceLimits(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled":                 "true",
 			"traefik.resources.limits.cpu":    "600m",
 			"traefik.resources.limits.memory": "600Mi",
@@ -180,8 +179,8 @@ func Test_checkResourceLimits(t *testing.T) {
 	)
 
 	var deployment v12.Deployment
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -190,15 +189,15 @@ func Test_checkResourceLimits(t *testing.T) {
 }
 
 func Test_checkDefaultResourceRequests(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled": "true",
 		}),
 	)
 
 	var deployment v12.Deployment
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -207,15 +206,15 @@ func Test_checkDefaultResourceRequests(t *testing.T) {
 }
 
 func Test_checkDefaultResourceLimits(t *testing.T) {
-	helmChartParser := common.NewHelmConfigParser(
-		common.NewHelmTest(t, helmChartRelativePath, map[string]string{
+	helmChartParser := NewHelmConfigParser(
+		NewHelmTest(t, helmChartRelativePath, map[string]string{
 			"traefik.enabled": "true",
 		}),
 	)
 
 	var deployment v12.Deployment
-	helmChartParser.Find(common.SearchResourceOption{
-		Name: "release-name-traefik",
+	helmChartParser.Find(SearchResourceOption{
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	}, &deployment)
 
@@ -223,25 +222,25 @@ func Test_checkDefaultResourceLimits(t *testing.T) {
 	require.Equal(t, "500Mi", deployment.Spec.Template.Spec.Containers[0].Resources.Limits.Memory().String())
 }
 
-var traefikResources = []common.SearchResourceOption{
+var traefikResources = []SearchResourceOption{
 	{
-		Name: "release-name-traefik",
+		Name: "pega-traefik",
 		Kind: "ConfigMap",
 	},
 	{
-		Name: "release-name-traefik",
+		Name: "pega-traefik",
 		Kind: "Deployment",
 	},
 	{
-		Name: "release-name-traefik",
+		Name: "pega-traefik",
 		Kind: "Service",
 	},
 	{
-		Name: "release-name-traefik-test",
+		Name: "pega-traefik-test",
 		Kind: "Pod",
 	},
 	{
-		Name: "release-name-traefik-test",
+		Name: "pega-traefik-test",
 		Kind: "ConfigMap",
 	},
 }
