+++ terratest/src/test/pega/constants.go
@@ -1,3 +1,5 @@
 package pega
 
 const PegaHelmChartPath = "../../../../charts/pega"
+const PegaHelmChartTestsPath = "../../../../terratest/src/test/pega"
+const PegaHelmRelease = "pega"
