+++ terratest/src/test/pega/pega-installer-status-rolebinding_test.go
@@ -1,0 +1,43 @@
+package pega
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/gruntwork-io/terratest/modules/helm"
+	"github.com/stretchr/testify/require"
+	k8srbac "k8s.io/api/rbac/v1"
+)
+
+func TestPegaInstallerStatusRoleBinding(t *testing.T) {
+	var supportedVendors = []string{"k8s", "openshift", "eks", "gke", "aks", "pks"}
+	var supportedOperations = []string{"install-deploy", "upgrade-deploy"}
+
+	helmChartPath, err := filepath.Abs(PegaHelmChartPath)
+	require.NoError(t, err)
+
+	for _, vendor := range supportedVendors {
+		for _, operation := range supportedOperations {
+			var options = &helm.Options{
+				SetValues: map[string]string{
+					"global.provider":        vendor,
+					"global.actions.execute": operation,
+				},
+			}
+			yamlContent := RenderTemplate(t, options, helmChartPath, []string{"charts/installer/templates/pega-installer-status-rolebinding.yaml"})
+			assertInstallerRoleBinding(t, yamlContent)
+		}
+	}
+}
+
+func assertInstallerRoleBinding(t *testing.T, roleBindingYaml string) {
+	var installerRoleBindingObj k8srbac.RoleBinding
+	helm.UnmarshalK8SYaml(t, roleBindingYaml, &installerRoleBindingObj)
+	require.Equal(t, installerRoleBindingObj.RoleRef.APIGroup, "rbac.authorization.k8s.io")
+	require.Equal(t, installerRoleBindingObj.RoleRef.Kind, "Role")
+	require.Equal(t, installerRoleBindingObj.RoleRef.Name, "jobs-reader")
+
+	require.Equal(t, installerRoleBindingObj.Subjects[0].Kind, "ServiceAccount")
+	require.Equal(t, installerRoleBindingObj.Subjects[0].Name, "default")
+	require.Equal(t, installerRoleBindingObj.Subjects[0].Namespace, "default")
+}
